+++ src/type_inference/resolve_bindings.ts
@@ -1,18 +1,22 @@
 import {
-  DeclaredType,
-  ResolvedType,
-  Type,
-} from '../types/type_inference/categories'
-import {
+  DeclaredBinding,
+  DeclaredTermBinding,
   LocalBinding,
   LocalTermBinding,
   LocalTypeBinding,
   TermBinding,
   TypeAssignment,
   TypeBinding,
+  isDeclaredBinding,
   isImportedBinding,
   isLocalBinding,
+  isTermBinding,
 } from '../types/analyze/bindings'
+import {
+  DeclaredType,
+  ResolvedType,
+  Type,
+} from '../types/type_inference/categories'
 import { ScopeWithErrors, TypedFileScope } from '../types/analyze/scopes'
 import {
   addErrorToScope,
@@ -44,7 +48,7 @@ const resolveBindingType = <
 >(
   getBindings: (fileScope: TypedFileScope) => StrongBinding<T>[],
   resolveLocalBindingType: (
-    binding: WeakBinding<T> & LocalBinding,
+    binding: WeakBinding<T> & (DeclaredBinding | LocalBinding),
     bindings: StrongBinding<T>[][],
   ) => W,
   buildAnswer: (result?: W) => W,
@@ -54,7 +58,10 @@ const resolveBindingType = <
   bindings: StrongBinding<T>[][],
   binding: WeakBinding<T>,
 ): [type: W, newScope: U, newFileScopes: TypedFileScope[]] => {
-  if (isLocalBinding(binding))
+  if (
+    isLocalBinding(binding) ||
+    (isTermBinding(binding) && isDeclaredBinding(binding))
+  )
     return [
       buildAnswer(resolveLocalBindingType(binding, bindings)),
       scope,
@@ -63,7 +70,7 @@ const resolveBindingType = <
 
   assert(
     isImportedBinding(binding),
-    'When a binding is not local it must be imported.',
+    'When a binding is not local or declared it must be imported.',
   )
 
   const importedBindingName = binding.originalName || binding.name
@@ -118,7 +125,7 @@ const buildAnswer = <T extends Type | DeclaredType>(answer?: T) =>
   answer || buildTemporaryTypeVariable()
 
 const resolveTermBindingTypeWithinScope = (
-  binding: LocalTermBinding,
+  binding: DeclaredTermBinding | LocalTermBinding,
   typeAssignments: TypeAssignment[][],
 ): ResolvedType[] => {
   const bindings = findBindings(binding.name, typeAssignments)
